refactors harvesting show files
